Split up Burgers Equations.cpp
